requirements.txt
python-telegram-bot==20.7
notion-client==2.2.1
APScheduler==3.10.4
hijri-converter==2.3.1
python-dotenv==1.0.0
pytz==2024.1
